validate provider function consistency
Provider functions must be pure, and return consistent results for the
same arguments. Store the results from all provider function calls in a
global cache, and ensure future calls always match prior results.

In the future we will store the data in the plan file, which will allow
us to verify the results again during apply.